changed method from public WatermarkBuilder opacity(int opacity) -> public WatermarkBuilder opacity(int opacity) in order to simplify api usage

+++ src/main/java/com/markit/api/AbstractWatermarkService.java
@@ -82,7 +82,7 @@ public abstract class AbstractWatermarkService<Service, WatermarkBuilder, TextBa
         return (WatermarkBuilder) this;
     }
 
-    public WatermarkBuilder opacity(float opacity) {
+    public WatermarkBuilder opacity(int opacity) {
         currentWatermark.setOpacity(opacity);
         return (WatermarkBuilder) this;
     }
